pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slingshot"
version = "0.1.0"
description = "Rapidly deploy and manage Cloudflare Workers with minimal hassle"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Mack", email = "mack@roamhq.io"}
]
keywords = ["cloudflare", "workers", "serverless", "deployment", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "click>=8.1.7",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",  # For beautiful CLI output
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "responses>=0.24.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
]

[project.scripts]
slingshot = "slingshot.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["slingshot*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=slingshot",
    "--cov-report=term-missing",
    "--cov-report=html",
]
